style(buffer-crypto): resolve new detekt alerts (line length, throws count, file function count)
Moves the four ECDH agreement helpers into HardwareKeyAgreement.android.kt
(HardwareKeys.android.kt was over the top-level function threshold), wraps the
new over-length lines, and suppresses ThrowsCount on agreeP256 matching the
jvmCommon deriveSharedSecretPlatform precedent. No behavior change.
